+++ e2e/flows/dist-outside-components.e2e.js
@@ -1,11 +1,17 @@
-import { expect } from 'chai';
+import chai, { expect } from 'chai';
 import fs from 'fs-extra';
 import path from 'path';
 import Helper from '../e2e-helper';
 
+chai.use(require('chai-fs'));
+
 describe('dists file are written outside the components dir', function () {
   this.timeout(0);
   const helper = new Helper();
+  const appJsFixture = `const barFoo = require('${helper.getRequireBitPath(
+    'bar',
+    'foo'
+  )}'); console.log(barFoo.default());`;
   let scopeWithCompiler;
   before(() => {
     helper.setNewLocalAndRemoteScopes();
@@ -53,8 +59,8 @@ describe('dists file are written outside the components dir', function () {
       helper.getClonedLocalScope(scopeWithCompiler);
       helper.importComponent('utils/is-type --dist');
 
-      const isStringFixture =
-        "import isType from 'bit/utils/is-type'; export default function isString() { return isType() +  ' and got is-string'; };";
+      const isStringFixture = `import isType from '${helper.getRequireBitPath('utils', 'is-type')}';
+ export default function isString() { return isType() +  ' and got is-string'; };`;
       helper.createComponent('utils', 'is-string.js', isStringFixture);
       helper.addComponent('utils/is-string.js');
       helper.commitAllComponents();
@@ -62,8 +68,8 @@ describe('dists file are written outside the components dir', function () {
       helper.getClonedLocalScope(scopeWithCompiler);
       helper.importComponent('utils/is-string --dist');
 
-      const fooBarFixture =
-        "import isString from 'bit/utils/is-string'; export default function foo() { return isString() + ' and got foo'; };";
+      const fooBarFixture = `import isString from '${helper.getRequireBitPath('utils', 'is-string')}';
+export default function foo() { return isString() + ' and got foo'; };`;
       helper.createComponentBarFoo(fooBarFixture);
       helper.addComponentBarFoo();
       helper.commitAllComponents();
@@ -74,21 +80,19 @@ describe('dists file are written outside the components dir', function () {
       helper.importComponent('bar/foo --dist');
     });
     it('should be able to require its direct dependency and print results from all dependencies', () => {
-      const appJsFixture = "const barFoo = require('bit/bar/foo'); console.log(barFoo.default());";
       fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
       const result = helper.runCmd('node app.js');
       expect(result.trim()).to.equal('got is-type and got is-string and got foo');
     });
     describe('"bit build" after updating the imported component', () => {
       before(() => {
-        const fooBarFixtureV2 =
-          "import isString from 'bit/utils/is-string'; export default function foo() { return isString() + ' and got foo v2'; };";
+        const fooBarFixtureV2 = `import isString from '${helper.getRequireBitPath('utils', 'is-string')}';
+export default function foo() { return isString() + ' and got foo v2'; };`;
         helper.createComponent('components/bar/foo', 'foo.js', fooBarFixtureV2); // update component
         helper.addRemoteEnvironment();
         helper.build();
       });
       it('should save the dists file in the same place "bit import" saved them', () => {
-        const appJsFixture = "const barFoo = require('bit/bar/foo'); console.log(barFoo.default());";
         fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
         const result = helper.runCmd('node app.js');
         expect(result.trim()).to.equal('got is-type and got is-string and got foo v2');
@@ -127,7 +131,6 @@ describe('dists file are written outside the components dir', function () {
       helper.importComponent('bar/foo --dist');
     });
     it('should be able to require its direct dependency and print results from all dependencies', () => {
-      const appJsFixture = "const barFoo = require('bit/bar/foo'); console.log(barFoo.default());";
       fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
       const result = helper.runCmd('node app.js');
       expect(result.trim()).to.equal('got is-type and got is-string and got foo');
@@ -211,6 +214,64 @@ describe('dists file are written outside the components dir', function () {
             path.join('my-dist', 'components', 'bar', 'foo', 'bar', 'src', 'foo.js')
           ); // both
         });
+      });
+    });
+  });
+  describe('when some dependencies have dists and some do not have', () => {
+    /**
+     * utils/is-type has dists
+     * utils/is-string doesn't have dists
+     * bar/foo has dists
+     */
+    before(() => {
+      helper.getClonedLocalScope(scopeWithCompiler);
+      helper.reInitRemoteScope();
+      const isTypeFixture = "export default function isType() { return 'got is-type'; };";
+      helper.createComponent('utils', 'is-type.js', isTypeFixture);
+      helper.addComponent('utils/is-type.js');
+      helper.commitAllComponents();
+      helper.exportAllComponents();
+
+      helper.reInitLocalScope();
+      helper.addRemoteScope();
+      helper.importComponent('utils/is-type --dist');
+
+      const isStringFixture = `const isType = require('${helper.getRequireBitPath(
+        'utils',
+        'is-type'
+      )}'); module.exports = function isString() { return isType.default() +  ' and got is-string'; };`;
+      helper.createComponent('utils', 'is-string.js', isStringFixture);
+      helper.addComponent('utils/is-string.js');
+      helper.commitAllComponents();
+      helper.exportAllComponents();
+      helper.getClonedLocalScope(scopeWithCompiler);
+      helper.importComponent('utils/is-string --dist');
+
+      const fooBarFixture = `import isString from '${helper.getRequireBitPath('utils', 'is-string')}';
+export default function foo() { return isString() + ' and got foo'; };`;
+      helper.createComponentBarFoo(fooBarFixture);
+      helper.addComponentBarFoo();
+      helper.commitAllComponents();
+      helper.exportAllComponents();
+      helper.reInitLocalScope();
+      helper.addRemoteScope();
+      helper.modifyFieldInBitJson('dist', { target: 'dist' });
+      helper.importComponent('bar/foo --dist');
+    });
+    it('should be able to require its direct dependency and print results from all dependencies', () => {
+      fs.outputFileSync(path.join(helper.localScopePath, 'app.js'), appJsFixture);
+      const result = helper.runCmd('node app.js');
+      expect(result.trim()).to.equal('got is-type and got is-string and got foo');
+    });
+    describe('"bit build" of a component with no compiler-id (when dists are outside the components dir)', () => {
+      before(() => {
+        fs.removeSync(path.join(helper.localScopePath, 'dist'));
+        helper.build('utils/is-string');
+      });
+      it('should save the source files as dists files', () => {
+        expect(
+          path.join(helper.localScopePath, 'dist', 'components', '.dependencies', 'utils', 'is-string')
+        ).to.be.a.path();
       });
     });
   });