+++ src/api/consumer/lib/import.js
@@ -22,7 +22,9 @@ export default (async function importAction({
   force,
   dist,
   conf,
+  installNpmPackages,
   withPackageJson,
+  saveDependenciesAsComponents,
   writeBitDependencies = false
 }: {
   ids: string,
@@ -34,9 +36,25 @@ export default (async function importAction({
   force: boolean,
   dist: boolean,
   conf: boolean,
+  installNpmPackages: boolean,
   withPackageJson: ?boolean,
+  saveDependenciesAsComponents: ?boolean,
   writeBitDependencies: ?boolean
 }): Promise<any> {
+  if (!withPackageJson) {
+    // if package.json is not written, it's impossible to install the packages and dependencies as npm packages
+    installNpmPackages = false;
+    saveDependenciesAsComponents = true;
+  }
+  if (!installNpmPackages) {
+    // if npm packages are not installed, don't install dependencies as npm packages
+    saveDependenciesAsComponents = true;
+  }
+  if (!saveDependenciesAsComponents) {
+    // according to the two 'if' statements above, installNpmPackages and withPackageJson must be true
+    writeBitDependencies = true;
+  }
+
   async function importEnvironment(consumer: Consumer): Promise<any> {
     loader.start(BEFORE_IMPORT_ENVIRONMENT);
 
@@ -74,7 +92,9 @@ export default (async function importAction({
     writeBitDependencies,
     force,
     dist,
-    conf
+    conf,
+    installNpmPackages,
+    saveDependenciesAsComponents
   );
   const bitIds = dependencies.map(R.path(['component', 'id']));
   const bitMap = await consumer.getBitMap();
@@ -91,7 +111,8 @@ export default (async function importAction({
   const warnings = await warnForPackageDependencies({
     dependencies: flattenDependencies(dependencies),
     envDependencies,
-    consumer
+    consumer,
+    installNpmPackages
   });
   return { dependencies, envDependencies, warnings };
 });
@@ -135,13 +156,13 @@ function compatibleWith(a: { [string]: string }, b: { [string]: string }): boole
 
 // TODO: refactor to better use of semver
 // TODO: move to bit-javascript
-const warnForPackageDependencies = ({ dependencies, consumer }): Promise<Object> => {
+const warnForPackageDependencies = ({ dependencies, consumer, installNpmPackages }): Promise<Object> => {
   const warnings = {
     notInPackageJson: [],
     notInNodeModules: [],
     notInBoth: []
   };
-
+  if (installNpmPackages) return warnings;
   const projectDir = consumer.getPath();
   const getPackageJson = (dir) => {
     try {