+++ src/cli/commands/public-cmds/import-cmd.js
@@ -21,7 +21,17 @@ export default class Import extends Command {
     ['f', 'force', 'ignore local changes'],
     ['', 'dist', 'write dist files (when exist) to the configured directory'],
     ['', 'conf', 'write the configuration file (bit.json)'],
-    ['', 'ignore-package-json', 'do not generate package.json for the imported component(s)']
+    ['', 'save-dependencies-as-components', 'save hub dependencies as bit components rather than npm packages'],
+    [
+      '',
+      'skip-npm-install',
+      'do not install packages of the imported components. (it automatically enables save-dependencies-as-components flag)'
+    ],
+    [
+      '',
+      'ignore-package-json',
+      'do not generate package.json for the imported component(s). (it automatically enables skip-npm-install and save-dependencies-as-components flags)'
+    ]
   ];
   loader = true;
   migration = true;
@@ -38,6 +48,8 @@ export default class Import extends Command {
       force = false,
       dist = false,
       conf = false,
+      skipNpmInstall = false,
+      saveDependenciesAsComponents = false,
       ignorePackageJson = false
     }: {
       tester?: boolean,
@@ -49,6 +61,8 @@ export default class Import extends Command {
       force?: boolean,
       dist?: boolean,
       conf?: boolean,
+      skipNpmInstall?: boolean,
+      saveDependenciesAsComponents?: boolean,
       ignorePackageJson?: boolean
     }
   ): Promise<any> {
@@ -56,7 +70,6 @@ export default class Import extends Command {
     if (tester && compiler) {
       throw new Error('you cant use tester and compiler flags combined');
     }
-
     return importAction({
       ids,
       tester,
@@ -67,7 +80,9 @@ export default class Import extends Command {
       force,
       dist,
       conf,
-      withPackageJson: !ignorePackageJson
+      installNpmPackages: !skipNpmInstall,
+      withPackageJson: !ignorePackageJson,
+      saveDependenciesAsComponents
     }).then(importResults => R.assoc('displayDependencies', displayDependencies, importResults));
   }