+++ src/scope/migrations/component/update-binding-prefix-to-new-default.js
@@ -1,0 +1,15 @@
+/** @flow */
+
+function updateBindingPrefixToNewDefault(componentModel: Object): Object {
+  const oldDefault = 'bit';
+  const newDefault = '@bit';
+  if (componentModel.bindingPrefix && componentModel.bindingPrefix === oldDefault) {
+    componentModel.bindingPrefix = newDefault;
+  }
+  return componentModel;
+}
+
+export default {
+  name: 'update bindingPrefix from "bit" to "@bit"',
+  migrate: updateBindingPrefixToNewDefault
+};